Move Reconciler into reconciler.go
